Update abilities, dashboard, and user authentication

+++ src/app/api/books/delete/route.ts
@@ -1,21 +1,64 @@
+import { defineAbilty } from "@/abilities";
 import { prisma } from "@/db";
+import { verify } from "@/lib/utils";
+import { User } from "@prisma/client";
 import { NextRequest, NextResponse } from "next/server";
 
 export const DELETE = async (req: NextRequest) => {
   try {
-    const json = (await req.json()) as { id: number };
     const token = req.cookies.get("token");
-    //TODO: check for authorization
+    const decoded = await verify<User>(token?.value ?? "");
+    if (!decoded) {
+      return NextResponse.json({
+        status: "error",
+        data: {
+          message: "Unauthorized",
+        },
+      });
+    }
 
-    const deleteBook = await prisma.book.delete({
+    const user = await prisma.user.findFirst({
       where: {
-        id: json.id,
+        email: decoded.email,
       },
     });
+
+    if (!user) {
+      return NextResponse.json({
+        status: "error",
+        data: {
+          message: "User not found",
+        },
+      });
+    }
+
+    const ablity = defineAbilty(user);
+
+    if (ablity.can("delete", "Book")) {
+      const { id } = (await req.json()) as { id: string };
+      if (!id)
+        return NextResponse.json({
+          status: "error",
+          data: {
+            message: "please send id of book ",
+          },
+        });
+      const deleteBook = await prisma.book.delete({
+        where: {
+          id: Number(id),
+        },
+      });
+      return NextResponse.json({
+        status: "success",
+        data: {
+          message: "the book has been deleted successfully",
+        },
+      });
+    }
     return NextResponse.json({
-      status: "success",
+      status: "error",
       data: {
-        message: "the book has been deleted successfully",
+        message: "you are not authorized to perform this action",
       },
     });
   } catch (err) {
